address comments to use dedicated hook for number of delegators query
